requirements.txt
RPi.GPIO==0.6.3
paho-mqtt==1.3.0
